refactor(input): fold cursor-icon bootstrap into the hyperlink feature
insert_initial_cursor_icon exists solely so hyperlink hover can mutate
CursorIcon without inserting it; the prerequisite joins its feature.
OrzmaBootstrapPlugin is gone.
